+++ sdk/core/src/CommonNetworkMember/BaseNetworkMember.ts
@@ -147,6 +147,8 @@ export abstract class BaseNetworkMember {
         registryUrl: registryUrl,
         metricsUrl: metricsUrl,
         component: eventComponent,
+        resolveLegacyElemLocally: options.otherOptions?.resolveLocallyElemMethod,
+        beforeDocumentLoader: options.otherOptions?.beforeDocumentLoader,
       },
       platformCryptographyTools,
     )