+++ sdk/expo/CHANGELOG.md
@@ -1,3 +1,7 @@
+# release 7.9.0 (2022-12-07)
+* added an option to pass custom before hook for document loader
+# release 7.8.0 (2022-12-07)
+* added a flag to resolve legacy elem locally
 # release 7.7.0 (2022-11-29)
 * add security fix for signIn passwordless flow
 # release 7.6.0 (2022-11-29)